+++ src/components/ScheduleCard.css
@@ -1,0 +1,64 @@
+.schedule-card {
+  display: flex;
+  align-items: center;
+  border: 1px solid #ddd;
+  border-radius: 8px;
+  padding: 10px;
+  background-color: #fff;
+  max-width: 100%;
+  cursor: grab;
+}
+
+.schedule-card-image {
+  width: 80px;
+  height: 80px;
+  border-radius: 5px;
+  margin-right: 10px;
+  object-fit: cover;
+}
+
+.schedule-card-details {
+  display: flex;
+  flex-direction: column;
+  flex: 2;
+}
+
+.schedule-card-details strong {
+  font-size: 14px;
+  margin-bottom: 5px;
+}
+
+.schedule-card-time-details {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  flex: 1;
+  font-size: 12px;
+  color: #555;
+}
+
+.schedule-card-actions {
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+  margin-left: 15px;
+}
+
+.schedule-card-actions button {
+  padding: 5px 10px;
+  font-size: 12px;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  cursor: pointer;
+  background: #f0f0f0;
+}
+
+.schedule-card-delete {
+  margin-left: auto;
+  padding: 5px 10px;
+  background: #ff4d4f;
+  color: #fff;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+}
